devtools/manifest: changing the logic for resolving manifest path
Previously, if the $VANADIUM_ROOT/.local_manifest existed, it would
override the value specified by the command-line -manifest flag. This
seems counterintuitive and likely to baffle users.

This CL changes the logic for resolving manifest paths as follows:

- by default all manifest flags are set to ""
- if the command-line -manifest flag is set, its value is respected
- otherwise, the logic uses $VANADIUM_ROOT/.local_manifest if it
  exists and uses the "default" manifest otherwise

NOTE: The new mechanism lends itself amenable to specifying a default
manifest for the v23 tool to use by creating
$VANADIUM_ROOT/.local_manifest as a symlink to for instance
$VANADIUM_ROOT/.manifest/v2/store.

Change-Id: Ifcd435226c6a7f83138a77e89acc7d876e82498c