devtools/internal/util: fix path to thirdparty tools.
Maybe I tested it wrong, but in some tests runtime.GOARCH and GOARCH env var are the same.
Maybe this is related to how v23 binary is built? I am not sure..
For now I am using a hard-coded "amd64" as runtime.GOARCH and added todo.

I also moved this function out of buildcop.go to misc.go

Change-Id: Ib348a75ec1003b1548963d0610dda757cd312eb9