devtools/v23: Detailed error information and fix up setup of the "arm" profile.
This commit:
(1) Provides more detail on failures. For example, if there are multiple
attempts to setup a profile, it prints out why each attempt failed.  In
verbose mode, it prints out the cause of failure for each command.

(2) Fixes an issue with the setup of the "arm" profile. For reasons I
haven't dug into yet, the crosstool-ng build creates
$VANADIUM_ROOT/environment/cout/xgcc/arm-unknown-linux-gnueabi with no
write permissions. As a result, after the setup is nearly complete and
all the heavy lifting of building the toolchain is over (8 hours after
the process began on the virtual machine on my laptop!), the profile
setup fails on writing the "action completed" file and restarts from
scratch. This commit introduces a change to hack around that permissions
issue.

Change-Id: If72001fea03ad464428988b477d33f10264aab24