prod-services-test: Test the HTTP service and then some.
Two basic changes:
(1) The identityd HTTP service is tested by fetching the public key
    it exports
(2) This public key is used to setup the principal that will be used
    by the "vrpc signature" commands that test prod services.

I have a related change to the "vrpc" tool to undo the "insecure by
default" scheme used for "vrpc signature" (see:
https://vanadium-review.googlesource.com/#/c/9040/). When that is in,
then (2) becomes useful as it ensures that the production services
are running with valid (not expired for example) blessings from
the appropriate identity provider.

If https://vanadium-review.googlesource.com/#/c/9040/ were submitted
before this change, then the prod services test would fail as the
"vrpc signature" invocations would refuse to talk to the services,
since they do not recognize the blessings presented by those services.

Change-Id: If62e205dcbb7ceb6eaa45213d5f8cd5b0f31e567