devtools/v23: set GOPATH and VDLPATH based on both the project manifest and the contents of the V23_ROOT
This CL avoids adding Go and VDL workspaces specified by the v23
configuration file into the GOPATH and VDLPATH respectively if these
workspaces are not specified in the project manifest or do not exist
locally. This is to avoid hard to debug errors where ghost (or
incorrect) workspaces are added to the GOPATH / VDLPATH used by the
v23 tool.

Change-Id: I9d3351bb90acdf94b0cb50a32b3ef52d8b4ee665