devtools: Replace x/devtools/internal/envutil with x/lib/envvar
A while ago I wrote envutil in the devtools repo, to make it
easier to deal with environment variables, and to make it easy
for the v23 tool to keep track of the mutations it had made.

Later on I noticed that there was similar code in x/ref in a few
places.

This CL removes envutil, and adds a new package envvar that may
be used by all other packages.  It behaves similarly to envutil,
and also to the code scattered around x/ref.

In a subsequent CL I'll add cmdline support for environment
variables using envvar.

Change-Id: I4547e97c0d95f48db0174c7be5827899d45c9654