TBR: v32/internal/test,vbinary: make vbinary work for the release snapshot binaries
The vbinary tool expects a certain path format (os_arch directory, full
timestamp dir names) and 'vanadium-binaries' storage bucket name.  This change
adds support to the vbinary tool to also work with the 'vanadium-release'
bucket, and the path structure used there.

The change in release.go adds the .done file to each directory, which is
expected by the vbinary tool.

Change-Id: I637053404cb9a2c0085a849c765e8cd1d95400ff