add Gradle installation and www tests
We're about to push our first Java tutorial and it depends on Gradle for
building. This CL downloads and installs Gradle 2.5 and takes care of
deleting it upon profile uninstallation. Removing any JDK installed by
`v23 profile install java` is left as a TODO.

This CL also adds test targets for the Java www test.

Change-Id: I7b1064f64f71c293f74bfe398621e871dcf8bc7f