devtools/v23: introducing support for dependent changes
This CL introduces support for dependent changes. To that end, it
introduces two new commands:

- "v23 cl new <name>" which should be used to create new (dependent)
  CLs. It is still possible to use "git checkout -b <name>" to create
  new CLs, but CLs created this way cannot be used as a base for a
  dependent CL.

- "v23 cl sync" which can be used to bring all CLs in the sequence of
  dependent CLs leading to the current branch up to date with the
  remote branch this CL pertains to. The use of the "v23 cl sync"
  command is optional (as it is invoked as part of the modified "v23
  cl mail" command).

This CL also changes the way the "v23 cl mail" command operates. In
particular, the command first invokes "v23 cl sync" for the current
branch to make sure that there are no conflicts between CLs in the
sequence of dependent CLs leading to the current branch. Next, the
command creates a review branch out with a commit for each CL in the
said sequence created by squashing all commits of the individual CLs.

This CL also changes the way commit messages (and other CL metadata
such as the "dependency path" introduce by this CL) are
persisted. Previously, the commit message has been recorded in a file
in the branch. This required extra work when creating the review
branch as this file should not be exported to Gerrit. The new approach
moves this file under the top-level .v23/<branch> directory. This was
necessary to make the above support for dependent branches to work and
it also simplifies the existing management of commit messages.

Change-Id: I0c6e62279695ac75c89faf3a023071c2d14b5856