Initial support for hooks in manifest
This adds support for specifying a number of <hooks> in a manifest
that will be run at v23 update time. Each hook is specified as a
path relative to one of the projects in the manifest as well as an
optional 'Interpreter' parameter to use an interpreter from the system
(like python or bash or whatnot) that is not in one of the specified
projects. Hooks have names and can be overridden or excluded in imports
just like projects or tools. Each <hook> can have any number of <args>
which are passed to the hook as arguments. Hooks are run in the order
that they are specified in the manifest to handle hooks that have
interdependencies.

Change-Id: I2af29a7668464cacaea69a5a989092004c1c8c87