v.io/jiri: fix cl mail to handle errors correctly both from gerrit and in parsing .netrc
For some time I've puzzled as to why the topic was not being correctly
set for my CLs. Turns out the problem was entirely of my own making,
but something changed in v23/jiri to break setting the topic. My
problem was that I have duplicated entries in .netrc for the
same host. The first one had the correct password, the second one
had an old password. It seems that the previos behaviour was to
use the first one encountered, but this changed to using the
second one. This would have been easy to detect if yiri reported
non-200 status codes as errors - i.e. if a request succeeds
(i.e. no error returned by the http client) then it may still
have failed - e.g. a 403 - as was the case here. I've made two
changes:

- check the response status code
- complain if duplicate hosts are found in .netrc

Change-Id: Ia6426619f59ffe606d550d4b73c7b09eb0c34f4e