jiri: Allow nested projects
This removes the restriction that projects cannot be nested.
First, it makes jiri keep looking for projects even if it finds a .jiri.
This does mean that it will keep walking more, but currently it does so
anyway for e.g. build directories under $JIRI_ROOT.

Second, it creates projects in an order consistent with their nesting.

Change-Id: I917a1ef291021a2537ce4a3ff69df722cd096abe