Do a git fetch + rebase <rev> instead of git pull origin master
This changes the behavior of 'jiri update' when a project specifies a specific
revision to better support tracking remote refs that are not in the remote's
master branch. Specifically it changes the update sequence from

git pull origin master
git reset --hard <revision>

to

git fetch origin
git rebase <refspec>
git reset --hard HEAD

where <refspec> is the Revision specified in the project or origin/master if
not specified.

Fixes vanadium/issues#821

Change-Id: I3f61b7c0c52913b626dfddadb0cab1a91def3781