project: Automatically add /.jiri/ to .git/info/exclude when creating project
Every repository that jiri creates will have untracked files in /.jiri/ that
jiri manages. Requiring every repo add this to their own .gitignore is
burdensome at least and in some cases not feasible when depending on an
arbitrary foreign repository. This adds /.jiri/ to the .git/info/exclude file
in repositories that jiri creates.

Change-Id: I104bef49384050b21b33301b34efc0b94027d0f5