cl: add a verify flag to control git push --(no-)verify
This adds a flag to 'jiri cl mail' that controls whether --no-verify
is sent to git push, which controls whether pre-push hooks are run.
The flag naming here is the same as 'git push' which maybe isn't the
most sensible but is at least consistent with using gerrit directly.

MultiPart: 1/2

Change-Id: Icc99e6b7e18c1801c7529797835b2359604d1a90