v.io/jiri: add policy driven flags for merging environment variables.
Environment variables may originate from the inherited environment
of the process or from profiles. The variables may need to be
combined (e.g. CGO_FLAGS or PATH) or chosen from one source
or other. No single policy has been found to work for the combination
of end user development and jenkins workflows for native and
cross compilation. The intent of this CL is to make it easy
to configure environment variables as needed in as obvious, yet
convenient way, as possible.

Also tidy up v23-profile env, list etc.etc

MultiPart: 1/11

Change-Id: I1342cd2a1c292107ba1027a5b2a2eb611c5f51ae