v.io/jiri/profiles: bug fix, ensure that a version is always specified.
In order to be able to unambiguosly remove profiles, its essential that
the manifest records the default version # that was used at install time.
This was ensured when using v23-profile install, but not when using
profiles.EnsureProfileIsInstalled. This CL fixes that, but this
requires fixing up the existing installations. To do so, I introduce
a 'sanitize' command that can perform various cleanups like this.
It is intended for use by end users and from presubmits.

Whilst doing this, I folded cleanup into the new sanitize command and
similarly folded the 'recreate' command into list --info.

MultiPart: 1/2

Change-Id: I78de04fe773c0c95a9a9873a42013f2c1941ce65