v.io/jiri: add the ability to store+use relative paths in profiles.
This makes it possible to more easily share manifest files when
debugging problems and also has the side effect of making the
manifest xml files more readable.

In order to stage releasing this to production, a new schema
version is introduce (4) which will trigger the new behaviour.
Once this code is checked in it can handle both version 3
(absolute paths) and 4 (relative paths), at which point we
deploy a 'cleanup' action to migrate from 3 to 4.

MultiPart: 1/2

Change-Id: Icc73da35bb343cc757d5b9ab0d2cae7372628ce3