jiri: fix jiri update --manifest=<manifest>
The jiri update command currently ignores the revisions specified in the
remote manifest: we indiscriminately fetch and use the master branch
revision from the remote server regardless if a revision was already
specified in the requested manifest.

It looks like this was broken by
https://vanadium-review.googlesource.com/#/c/16663.

The fix is to only use the remote server's revision when the requested
revision is HEAD (otherwise, respect the revision requested).

Change-Id: I2bd1a49e2ea974f6fc8c0efc1530d89e17c669d3