jiri/runutil: fix a couple go test --race failures
The race detector complains about data races in accessing the out buffer
in start_test.  In TestStartCommandOK, we can rewrite the test to avoid
the need for checking the output of a child command that's still
executing (a cop-out, but didn't want to get into making the buffering
of output thread-safe just now). In TestStartCommandWithOptsOK, we can
wait for the command to complete before checking the output.

Change-Id: I2b9424484500189ced2c5e2854241403790024e4