When updating projects, set origin to the remote in the manifest.
Currently, changing the remote in the manifest after a project has been
created has no effect.  This CL will set the origin to to the manifest
remote.  It will error if the project does not have a remote listed in
the manifest.

It also renames RepoName to RemoteUrl, to better match its
functionality.

Change-Id: Ia114b14eb2c10861205924aa02dd2b088cec15fb