RelativePaths are always relative to JIRI_ROOT.
RelativePaths used to be more general in that they could be made
relative to anything.  The common use case was to make then relative to
JIRI_ROOT.

Now, RelativePaths must be relative to JIRI_ROOT.  When Expanding the
RelativePath to a full absolute path, a "jirix" must be passed in to
provide the value of JIRI_ROOT.

RelativePaths moved from profiles package to jiri package.

Also add jiri project to apiCheckProjects.

MultiPart: 1/2
Change-Id: Iba4483e248d8eed758486ee061e3eac2960fcbd4