jiri: Remove jiritest.NewX_DeprecatedEnv
In most of the tests, I replaced NewX_DeprecatedEnv with
NewFakeJiriRoot.  That broke tests that used profiles, since the
fake JIRI_ROOT doesn't contain the third_party project, which is
for various profiles-related setup.  The fix for those cases is
to pass -skip-profiles; that also makes the tests a bit more
hermetic, and run faster.

MultiPart: 1/2

Change-Id: Ia9de91a19b586abd59d4e65c036fb8fa8e826072