Correct the ordering of operations
A check for one operation creating a repo at a path that is a prefix of
another create operation was added to the Less implementation on
operations. However, is-a-prefix-of is not a total ordering, and so Less
then also failed to be. This was causing the erratic and incorrect
sorting behavior.

This patch drops the dependence on package name and instead simply sorts
on the lexigraphical ordering of operation type followed by package path.

Change-Id: I84812358bf2b131a062c2fe67a2dad73bb7d9a8c