TBR jiri: Create and populate the .jiri_root dir
Our plans are to move to a new world order where there is a
.jiri_root directory that contains all the
root-level (i.e. non-project-level) jiri metadata.  This is a
step in that direction.

This CL creates the .jiri_root directory when "jiri update" is
run.  In our final state we want the bootstrap script to do this,
but we can't force everyone to re-run the bootstrap, so this is
part of the the transition logic.

We also populate this with two directories:
  root/.jiri_root/bin            # previously root/devtools/bin
  root/.jiri_root/update_history # previously root/.update_history

It's important to move the bin directory now, since we don't want
our shim script to change very often (or at all), so we need the
new location.  The "jiri update" logic now populates the new bin
directory, and we have transitional logic that backs up an
existing devtools/bin into .jiri_root/bin.BACKUP, and drops a
symlink from devtools/bin to .jiri_root/bin.

Also added the shim script, which is very straightforward.

MultiPart: 1/2
Change-Id: Ia90c337d2ec363d12033599fd2b3e303cd518a3a