jiri: Allow deletion+creation of projects with the same path during update.
We keep track of which project directories are deleted during the course
of an update.  This way, when testing if a creation is possible, we can
detect that the destination will be empty even if it isn't right now.

This delete+create case is important for project renames, which are
treated like delete+create.

Change-Id: Ia425417bd09d5aa9b604306c7431c3b62bae543d