Allow deletion of the jiri project.
This is preventing us from moving off of git-mirror, because that will
require a deletion+creation of the jiri project.

Note that the intention of this check is already broken, because we do
allow the deletion of devtools and x/lib, so it's been possible to break
the jiri tool if you update with a bad manifest.

Change-Id: Ib862d27cf6921b12ab3fb799dba93f479b6ad986