jiri: Transition away from local/remote snapshots.
The "jiri snapshot" behavior had the notion of "local" and "remote"
snapshots, which were saved in specific locations, and committed to git
history.  Remote snapshots were also pushed upstream.

This CL adds two new flags:
* -dir specifies the location to save the snapshot file in.
* -push-remote specifies that the snapshot should be committed and
  pushed upstream

The -remote flag is kept around for backwards compatibility, but all
uses of it should be replaced with a combination of -dir and -sync.

New tests are added that the functionality of the new flags.

MultiPart: 1/2

Change-Id: I97f33e055687eae5a8fec4e336e18ae9871fc4e5