v.io/jiri: runutil environment tidyup and move gitutil creation out of context.
- the previous implementation of runutil/Sequence, or rather
the underlying executor would always merge the operating
system environment with the environment carried by the
sequence. I found this surprising and hard to debug so
I've modified the Sequence code to behave as follows:
   1. if it is created with a nil environment it will
      use the operating system environment.
   2. Sequence.Env overrides all previous environment
      settings
   3. The new Sequence.MergeEnv merges the supplied
      environment with that already carried by the
      sequence.
This new behaviour seems more consistent and predictable.
In particular, jiri/X.NewSeq() creates a new sequence
with the environment already in use by jiri.X and as such
is predicatable also. In nearly all cases, jiri.X is created
with the operating system environment so the behaviour before
and after this change will be the same except for cases where
the code uses os.Setenv between creating the jiri.X and calling
Sequence.Run,Last etc. This is mostly tests.

- this change also moves gitutil.Git creation out of the context
to simplify dependency management since it's currently very easy
to create dependency cycles.

MultiPart: 1/2

Change-Id: Ia1031b54d7841010ffe1775e2873a2a911ce8aca