jiri: Remove Hosts and make gerrithost and githooks attributes on Project.
Previously, manifests had a Hosts tag that could contain a "gerrit" host
and a "git" host.  This assumed that all projects in the manifest used
the same git and gerrit hosts, which was a bad assumption, and now that
we have remote imports it doesn't make any sense.

The "gerrit" host tag is gone, and "gerrithost" is now an attribute on
each project.

The "git" host was removed entirely, as it can be inferred from the
project.repo.

Lastly, the "git" host was allowed to contain githooks which would be
installed for all projects in the manifest.  This has been changed so
that project tags contain a "githooks" attribute that points to a
directory of githooks to install for that project.

The necessary changes to the manifest repo have already been checked in.

MultiPart: 1/2

Change-Id: I230eb74390b83a3499796d6fccfda89c117cfac0