TBR: jiri: Remove <hook> tags.  "runhook" is now an attr on <project>.
Previously <hooks> was a child of <manifest>, and each <hook> specified
which project it belonged to.

Now, "runhook" is an attribute of <project>.  It points to a single
script that will get run when the project is created, updated, or moved.
The path is interpreted relative to the root.

We run all project hooks after updating each project, but before
building the tools.

Change-Id: I96c1266a301915b172a17f122ae1fd42d52afa9b