jiri: Don't try to talk to Gerrit REST API if gerrit host is not http(s).
When setting the CL topic, jiri makes an HTTP request to a Gerrit REST
API endpoint.

If the gerrit host url does not have an "http(s)" scheme, then we should
not try to talk to the host over HTTP.  Instead, we give an error to the
user telling them to use a host with "https", or run with
--set-topic=false.

In order to do this, I changed the CLOpts and Gerrit structs so that
"host" is a *url.URL, so we don't need to re-parse the url in different
places.

Fixes https://github.com/vanadium/issues/issues/915

MultiPart: 1/2

Change-Id: I8f0c2015e48684b79809ce8a11d12a14683cc4ee