jiri: Close file after Copy()ing.
Previously we were not calling file.Close() after copying to the file,
causing later operations (like Chmod) to fail because the file was still
open.

There is only one call to seq.Copy(), so I added the file.Close() there.

I also changed the signature of seq.Copy() to take an io.Writer instead
of an os.File.  This makes the signature match io.Copy, and makes it
clear that seq.Copy will not close the file for you.

Change-Id: I4e85ddc59c52200f85afae29ea0f80e9e01ab3b0