TBR: jiri: Don't try to load the .jiri_manifest if we haven't upgraded yet.
Our last "hotfix" in v.io/c/19373 only works if the user has upgraded to
the new-style manifests.  If they have not upgraded, it errors when
reading the .jiri_manifest file.

This CL fixes that by calling LocalProjects if we are in the old world,
and loadManifestFile if we are in the new world.

Change-Id: I62ed03e59e63912e6845627b2a2a1edafb650b0e