TBR: jiri: Fix two occurances of LoadManifest.
Under the new jiri import logic, calling LoadManifest will perform git
operations in the manifest directory.  This causes problems for things
that run in parallel, like "jiri go test".

The fix in profilereader should be OK, since it only needs a list of
projects, it can call LocalProjects.

The fix in paths.go is a total HACK, but it was already a total HACK.

Change-Id: Idf7b784781e3f7ffbd2028c620a8dd6999d1a9d3