profiles/profileutil: Set the execute bit on directories.
As the comments suggest, it is possible for a file to appear before the
directory when walking through the zip file contents. If the file does
not have the execute bit set then the directory was being created
without it and file creation inside that directory was failing with
"permission denied".

The fix is to always set the execute bit on the directory being
created.

Change-Id: Ia781cba882a4e259bc670ff7602187ee2e4ad7f7