jiri: Make sure that project Revision exists on RemoteBranch
Projects can specify both a Revsion and a RemoteBranch.  Previously, it
was possible to specify a Revision that does not occur on the
RemoteBranch.  In that case, we would reset to the Revision and ignore
the RemoteBranch.

This CL makes it impossible to create a project with conflicting
Revision and RemoteBranch.

Note that if RemoteBranch is not specified, it will default to "master".
 Hence, if you specify a Revision that does not occur on the master
 branch, you must also specify a RemoteBranch that does contain that
 particular revision.

Change-Id: I249ddeb5fef4d5b756e1fc7c0b069ca08df6d062