jiri: The jiri script must put $JIRI_ROOT/.jiri_root/bin in PATH before invoking the jiri binary.
Our setup instructions tell the user to add
$JIRI_ROOT/.jiri_root/scripts to their PATH.

The idea is that when they call "jiri", the script in
.jiri_root/scripts/jiri will be invoked, and that script will call
.jiri_root/bin/jiri.

However, when running a command like "jiri go install blah", the jiri
binary calls exec.LookPath to find the location of the "jiri-go" binary,
but that binary will not be in the user's PATH, and the user gets an
error "unknown command go".

Note that this bug can actually be much worse if the user has an old
Vanadium checkout and still has /old-jiri-root/devtools/bin in their
path.  In that case when they run "jiri go install blah", the jiri
binary in their current JIRI_ROOT will end up invoking the old jiri-go
binary in /old-jiri-root/devtools/bin.  In that case (which Aaron ran
into), it's very hard to notice that you are using an old jiri-go
binary.

Change-Id: Ia7f7e23988e66d4b9d649192c546d3403c19bab5