TBR: jiri: Copy scripts/jiri to JIRI_ROOT/.jiri_root/scripts/jiri when updating
It should be rare that we need to update the jiri script, but it
happens.  See v.io/c/20086

When we do update the jiri script, it should be installed in
JIRI_ROOT/.jiri_root/scripts/jiri during the next "jiri update".  This
CL does that.

This is the same CL as v.io/c/20112, which was reverted, but with an
extra call to MkdirAll.

Change-Id: Ied4e254b31a56f7ed9a4e6a83a3aadf551315072