Revert "jiri: Make sure that project Revision exists on RemoteBranch"
This reverts commit b2ad724c442930c6ade8dd29e5aff6dfe100e9f4.

The goal of this feature was to prevent users from specifying an
"invalid" configuration with both a revision and remotebranch, where the
revision does not occur on the remotebranch.

However, the old semantics said that "if revision is specified, 
remotebranch should be ignored", and so specifying a conflicting
revision and remotebranch isn't actually "invalid", just possibly
strange.

We considered combining the revision and remotebranch attributes into a 
single attribute, but that had two problems:
1) What do we call this attribute?  Typing "revision=my-branch" seems
   strange, and I can't think of anything feels more natural.
2) The jiri code treats revisions and remote branches differently, so we
   would end up having to sniff the attribute to detect whether it is a 
   revision or branch name (or tag), and I don't want to do that.

Instead, we'll revert the previous CL that caused conflicting 
revision/remotebranch to be an error.

In a subsequent CL, I'll add a line of documentation that says:
"If revision is specified, remotebranch is ignored."

Change-Id: I5c5d1f3b201ee0bacffcb22aa3fb4863af0f946b