jiri: Be consistent about absolute/relative paths.
The new policy is that Projects in memory should have absolute paths,
but files that we write (manifests and metadata) should use relative
paths.

This policy applies to all Project properties that are paths.  Currently
this includes Path, RunHook, and GitHooks properties.

To help enforce this policy, there are two new methods on Project:
relativizePaths and absolutizePaths.  relativizePaths is called when
writing a project to file, and absolutizePaths is called when reading a
project.

Note that unlike ProjectFromFile, the ManifestFromFile function does not
convert project paths to absolute paths.  This is because when loading a
manifest through a remote import, its common to specify a root directory
different from the jiri root.  The usual way to load a manifest is
though LoadManifest, which does convert project paths to absolute paths,
and uses the correct root directory.

This CL also removes the "root" property from operations, and we stop
passing "root" around during computeOperations.  This was always the
empty string, so it never mattered.

Change-Id: I48b03ff7730b1e86b90717330c632ef9d02e5e10