jiri: When importing a project with a root, prepend the root to the project name.
Consider two manifest, "foo" and "bar", both of which import manifest
"baz".  It should be possible to import both "foo" and "bar" manifests
under different roots, and end up with two copies of the projects in
"baz", under different roots.

This is not currently possible, because jiri enforces that project Keys
be unique, and the project Key contains only the name and remote.

This CL changes jiri import semantics such that when importing a
manifest with a root, all of the projects listed in that manifest will
have the root prepended to their names.

Thus, in the example above, the projects in the "baz" manifest will end
up with different names, and will be treated as unique by jiri.

Change-Id: I5c413bc4a249fdb8f318a2ba645fcd3403829a32