Make api-fix check the apis of all tracked files, not just those with changes.
"jiri api fix" was only regenerating the .api fils for files that
differed from those in master.  This led to some strange behavior:

* A file had to be committed before "jiri api fix" would generate the
  .api file for it.

* If a file changed and .api file was generated, but then the original
  file was reverted to be the same as on master, then "jiri api fix"
  would never update the .api file back to what it was.

This CL makes "jiri api fix" generate .api files for all tracked files
in the project, not just those that differ from master.

As a side-effect, it found some files in the jiri repo that expose a
public api that did not have a corresponding .api file.

MultiPart: 1/2
Change-Id: I7f30114943d9491ca318daa5fde1c469da3393e1